requirements.txt
fastapi
uvicorn
pydantic

langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-groq

chromadb
sentence-transformers
rank-bm25

python-dotenv